test: 受注一覧 表示項目設定の E2E を admin-order.spec.ts へ統合 (#6816)
新設の admin-order-display-setting.spec.ts は e2e-test.yml の matrix に
未登録で CI 実行されないため、受注一覧画面のテストである admin-order.spec.ts
（matrix 登録済みの admin-order suite）へ statement describe として統合する。
goOrderList / adminRoute は既存ヘルパーを再利用し、afterEach は当 describe に
スコープを閉じて他テストへ影響しないようにする。page object は維持。
